Add user list API endpoint
A new API endpoint '/users' was added to list all users. A new UserController was created and added to the api routes to handle this function. The UserController's 'index' method returns an array of all users, including their id, name and email. This update will help in managing user data more efficiently.