Add new routes and views for user and home pages
Refactored web routes to use HomeController instead of closure for the root route. HomeController now has methods for both "users" and "home", each returning a corresponding inertia.js view. New vue.js components: Users.vue and Home.vue have been added for the respective pages. Users.vue fetches and lists users from the API, with the ability to log in as a user by clicking on a row. Home.vue receives a 'user' prop and shows a greeting message with the user name.