Refactor Users.vue to use shared data
In this commit, the users data loading process in Users.vue was changed from an async loading function to using shared data from the Inertia middleware. This change was made to decrease the loading time and simplify the code by reducing the number of async requests. Now, the users data is shared directly via the Inertia's share method in HandleInertiaRequests middleware.