Add many-to-many relationship between User and ChatRoom
This change adds a many-to-many relation (Users can belong to many ChatRooms and vice versa) to the User and ChatRoom models by introducing a pivot table. This was needed to allow multiple users to be part of a single chat room. Adding this relationship will help in querying chat rooms a user belongs to and users in a particular chat room. A new migration file for the 'chat_room_user' pivot table has been created and the appropriate relation methods have been added in both User and ChatRoom models.