Add chatroom functionality and messages backend
A new MessageController has been added to provide backend support for message posting. The Vue-based ChatWidget component is now dynamically using an actual chat room data and user list fetched from the server. The `store` method in MessageController now enables storing sent messages in the database. Messages and room related information are now passed to the ChatWidget.vue from page props and used accordingly. The chat widget can now send POST requests to the `/messages` route, containing the body of the message and room's ID to continue with the storage of the message in the database.