Add new pages and navigation
This commit introduces two additional pages (Page1 and Page2) to the application. As part of these changes, new routes were also added to the web.php file, and respective methods were created in the HomeController. Moreover, a new Navigation component was created to facilitate navigation between pages. It's added to all the pages including the Home page. This enhancement was done to improve the user's navigational experience across the website.