Update ChatWidget CSS for improved room selector scrollability
Removed 'flex-grow: 1' and 'display: flex' from the .room-selector and added 'max-height: 350px' and 'overflow-y: auto' to improve scrollability of the room selector. This change increases usability when the room list exceeds the height of the chat widget. Also moved the 'margin-bottom: 5px;' from .users-container .room-btn to .room-btn for a better structuring of the CSS code.