Add toast notification for new messages
A toast notification feature has been added to the ChatWidget component in this commit.
This feature uses the "vue-toastification" package.
When a new chat message is received, a toast notification will be displayed announcing the new message and the room in which it was sent. By clicking on the toast, the user is directed to that chat room and the chat widget is expanded to display messages. This feature improves user interaction with chat rooms by keeping them aware of new messages.