Refactor user data seeding process
Removed individual user seeding process from DatabaseSeeder and moved it to the newly created UserSeeder class. This enhances code organization and scalability. Now we have separate UserSeeder for the creation of users which can be extended easily in the future for more complex user seeding operations. It also ensures only relevant user seeding logic is kept in UserSeeder, making it easy to manage.