Update README to reflect project specifics
README.md was significantly overhauled to accurately reflect the current project. In the previous version, the file contained information and instructions pertinent to the Laravel framework. This has been replaced with details and installation instructions specific to the Laravel Vue Chat application. This modification will provide clear and appropriate instructions to users and potential contributors.