Add addUser method to ChatRoom Controller
A new method, `addUser`, was added to the `ChatRoomController`. This method allows for adding new users to a given chatroom, by utilizing the `ChatRoomAddUsersRequest` input request. The related route was also added in the web.php file. This change enables more dynamic and flexible chatroom management by enabling the addition of users to specific chat rooms.