Add AddUserToRoom.vue component
Implemented a new Vue component, AddUserToRoom.vue, designed to facilitate the option of adding users to a chat room. The component includes a button for adding users and a dropdown for selecting the particular user. Moreover, it handles the API calls to add selected users to the chat room, making sure that the same user cannot be added twice. The users already present in the room are filtered out from the dropdown list.