Add clickOutside functionality to user dropdown
A Vueuse library function "onClickOutside" was imported and implemented in the AddUserToRoom.vue component. Its purpose is to detect when a user clicks outside of the user dropdown, in which case the dropdown menu will close. This improves the UI experience by providing a more typical and intuitive way for users to exit the dropdown menu. This feature is crucial for better app navigation and reducing confusion. Additionally, a new ref "dropdown" was created to target the dropdown element properly.