Add feature to invite users for chat in widgets
Added the AddUserToRoom component to the ChatWidget. This feature allows users to invite other users to their chatroom. The list of available users and the current room is passed as props to the AddUserToRoom component. When a user is added, it triggers the @add-user event, which updates the currentRoom's user list.