Change broadcast and queue drivers for better performance
This commit updates the values of BROADCAST_DRIVER and QUEUE_CONNECTION in the .env.example file. The BROADCAST_DRIVER is changed from 'log' to 'pusher' and QUEUE_CONNECTION from 'sync' to 'database'. This change is made to enhance application's performance by utilizing the Pusher broadcasting services and database queue connection.